chore: fix lint and tests
