+++ api/src/Errors/InvalidUrlError.ts
@@ -1,5 +1,5 @@
 export class InvalidUrlError extends Error {
-    constructor(url: string) {
-        super(`Url: ${url} is invalid.`);
-    }
+  constructor(url: string) {
+    super(`Url: ${url} is invalid.`);
+  }
 }
